requirements.txt
tensorflow
tensorflow-datasets
tensorflow.keras
flask
opencv-python
pandas
matplotlib
scikit-learn
Pillow
gunicorn